refactor(core): Split out database types
