+++ packages/cli/src/controllers/__tests__/me.controller.test.ts
@@ -14,11 +14,11 @@ import { BadRequestError } from '@/errors/response-errors/bad-request.error';
 import { InvalidMfaCodeError } from '@/errors/response-errors/invalid-mfa-code.error';
 import { EventService } from '@/events/event.service';
 import { ExternalHooks } from '@/external-hooks';
-import type { PublicUser } from '@/interfaces';
 import { License } from '@/license';
 import { MfaService } from '@/mfa/mfa.service';
 import type { AuthenticatedRequest, MeRequest } from '@/requests';
 import { UserService } from '@/services/user.service';
+import type { PublicUser } from '@/types-db';
 import { mockInstance } from '@test/mocking';
 import { badPasswords } from '@test/test-data';
