+++ packages/cli/src/controllers/oauth/abstract-oauth.controller.ts
@@ -15,9 +15,9 @@ import { AuthError } from '@/errors/response-errors/auth.error';
 import { BadRequestError } from '@/errors/response-errors/bad-request.error';
 import { NotFoundError } from '@/errors/response-errors/not-found.error';
 import { ExternalHooks } from '@/external-hooks';
-import type { ICredentialsDb } from '@/interfaces';
 import type { AuthenticatedRequest, OAuthRequest } from '@/requests';
 import { UrlService } from '@/services/url.service';
+import type { ICredentialsDb } from '@/types-db';
 import * as WorkflowExecuteAdditionalData from '@/workflow-execute-additional-data';
 
 type CsrfStateParam = {
