+++ packages/cli/src/databases/entities/workflow-entity.ts
@@ -12,7 +12,7 @@ import { Length } from 'class-validator';
 import { IConnections, IDataObject, IWorkflowSettings, WorkflowFEMeta } from 'n8n-workflow';
 import type { IBinaryKeyData, INode, IPairedItemData } from 'n8n-workflow';
 
-import type { IWorkflowDb } from '@/interfaces';
+import type { IWorkflowDb } from '@/types-db';
 
 import { WithTimestampsAndStringId, dbType, jsonColumnType } from './abstract-entity';
 import { type Folder } from './folder';
