+++ packages/cli/src/interfaces.ts
@@ -1,10 +1,9 @@
-import type { AssignableRole, GlobalRole, Scope } from '@n8n/permissions';
+import type { AssignableRole } from '@n8n/permissions';
 import type { Application } from 'express';
 import type {
 	ExecutionError,
 	ICredentialDataDecryptedObject,
 	ICredentialsDecrypted,
-	ICredentialsEncrypted,
 	IDeferredPromise,
 	IExecuteResponsePromiseData,
 	IRun,
@@ -15,8 +14,6 @@ import type {
 	WorkflowExecuteMode,
 	ExecutionStatus,
 	ExecutionSummary,
-	FeatureFlags,
-	IUserSettings,
 	IWorkflowExecutionDataProcess,
 	DeduplicationMode,
 	DeduplicationItemTypes,
@@ -24,15 +21,8 @@ import type {
 import type PCancelable from 'p-cancelable';
 
 import type { ActiveWorkflowManager } from '@/active-workflow-manager';
-import type { AnnotationTagEntity } from '@/databases/entities/annotation-tag-entity.ee';
-import type { AuthProviderType } from '@/databases/entities/auth-identity';
-import type { SharedCredentials } from '@/databases/entities/shared-credentials';
-import type { TagEntity } from '@/databases/entities/tag-entity';
-import type { User } from '@/databases/entities/user';
-
-import type { Folder } from './databases/entities/folder';
-import type { ExternalHooks } from './external-hooks';
-import type { WorkflowWithSharingsAndCredentials } from './workflows/workflows.types';
+import type { ExternalHooks } from '@/external-hooks';
+import type { ICredentialsBase, IExecutionBase, ITagBase } from '@/types-db';
 
 export interface ICredentialsTypeData {
 	[key: string]: CredentialLoadingDetails;
@@ -60,38 +50,14 @@ export interface IProcessedDataEntries {
 //               tags
 // ----------------------------------
 
-export interface ITagBase {
-	id: string;
-	name: string;
-}
-
 export interface ITagToImport extends ITagBase {
 	createdAt?: string;
 	updatedAt?: string;
 }
 
-export type UsageCount = {
-	usageCount: number;
-};
-
-export type ITagDb = Pick<TagEntity, 'id' | 'name' | 'createdAt' | 'updatedAt'>;
-
-export type ITagWithCountDb = ITagDb & UsageCount;
-
-export type IAnnotationTagDb = Pick<AnnotationTagEntity, 'id' | 'name' | 'createdAt' | 'updatedAt'>;
-
-export type IAnnotationTagWithCountDb = IAnnotationTagDb & UsageCount;
-
 // ----------------------------------
 //            workflows
 // ----------------------------------
-
-// Almost identical to editor-ui.Interfaces.ts
-export interface IWorkflowDb extends IWorkflowBase {
-	triggerCount: number;
-	tags?: TagEntity[];
-	parentFolder?: Folder | null;
-}
 
 export interface IWorkflowResponse extends IWorkflowBase {
 	id: string;
@@ -110,40 +76,11 @@ export interface IWorkflowToImport
 //            credentials
 // ----------------------------------
 
-export interface ICredentialsBase {
-	createdAt: Date;
-	updatedAt: Date;
-}
-
-export interface ICredentialsDb extends ICredentialsBase, ICredentialsEncrypted {
-	id: string;
-	name: string;
-	shared?: SharedCredentials[];
-}
-
 export type ICredentialsDecryptedDb = ICredentialsBase & ICredentialsDecrypted;
 
 export type ICredentialsDecryptedResponse = ICredentialsDecryptedDb;
 
 export type SaveExecutionDataType = 'all' | 'none';
-
-export interface IExecutionBase {
-	id: string;
-	mode: WorkflowExecuteMode;
-	createdAt: Date; // set by DB
-	startedAt: Date;
-	stoppedAt?: Date; // empty value means execution is still running
-	workflowId: string;
-
-	/**
-	 * @deprecated Use `status` instead
-	 */
-	finished: boolean;
-	retryOf?: string; // If it is a retry, the id of the execution it is a retry of.
-	retrySuccessId?: string; // If it failed and a retry did succeed. The id of the successful retry.
-	status: ExecutionStatus;
-	waitTill?: Date | null;
-}
 
 // Data in regular format with references
 export interface IExecutionDb extends IExecutionBase {
@@ -156,18 +93,6 @@ export type CreateExecutionPayload = Omit<IExecutionDb, 'id' | 'createdAt' | 'st
 
 /** Payload for updating an execution. */
 export type UpdateExecutionPayload = Omit<IExecutionDb, 'id' | 'createdAt'>;
-
-export interface IExecutionResponse extends IExecutionBase {
-	id: string;
-	data: IRunExecutionData;
-	retryOf?: string;
-	retrySuccessId?: string;
-	workflowData: IWorkflowBase | WorkflowWithSharingsAndCredentials;
-	customData: Record<string, string>;
-	annotation: {
-		tags: ITagBase[];
-	};
-}
 
 // Flatted data to save memory when saving in database or transferring
 // via REST API
@@ -219,16 +144,6 @@ export interface IExecutingWorkflowData {
 	responsePromise?: IDeferredPromise<IExecuteResponsePromiseData>;
 	workflowExecution?: PCancelable<IRun>;
 	status: ExecutionStatus;
-}
-
-export interface IPersonalizationSurveyAnswers {
-	email: string | null;
-	codingSkill: string | null;
-	companyIndustry: string[];
-	companySize: string | null;
-	otherCompanyIndustry: string | null;
-	otherWorkArea: string | null;
-	workArea: string[] | string | null;
 }
 
 export interface IActiveDirectorySettings {
@@ -323,26 +238,6 @@ export interface ILicensePostResponse extends ILicenseReadResponse {
 	managementToken: string;
 }
 
-export interface PublicUser {
-	id: string;
-	email?: string;
-	firstName?: string;
-	lastName?: string;
-	personalizationAnswers?: IPersonalizationSurveyAnswers | null;
-	password?: string;
-	passwordResetToken?: string;
-	createdAt: Date;
-	isPending: boolean;
-	role?: GlobalRole;
-	globalScopes?: Scope[];
-	signInType: AuthProviderType;
-	disabled: boolean;
-	settings?: IUserSettings | null;
-	inviteAcceptUrl?: string;
-	isOwner?: boolean;
-	featureFlags?: FeatureFlags;
-}
-
 export interface Invitation {
 	email: string;
 	role: AssignableRole;
@@ -354,5 +249,3 @@ export interface N8nApp {
 	externalHooks: ExternalHooks;
 	activeWorkflowManager: ActiveWorkflowManager;
 }
-
-export type UserSettings = Pick<User, 'id' | 'settings'>;
