+++ packages/cli/src/modules/insights/database/entities/insights-raw.ts
@@ -3,8 +3,9 @@ import { Container } from '@n8n/di';
 import { BaseEntity, Column, Entity, PrimaryGeneratedColumn } from '@n8n/typeorm';
 import { UnexpectedError } from 'n8n-workflow';
 
+import { datetimeColumnType } from '@/databases/entities/abstract-entity';
+
 import { isValidTypeNumber, NumberToType, TypeToNumber } from './insights-shared';
-import { datetimeColumnType } from '../../../../databases/entities/abstract-entity';
 
 export const { type: dbType } = Container.get(GlobalConfig).database;
