+++ packages/cli/src/public-api/v1/handlers/credentials/credentials.service.ts
@@ -16,8 +16,8 @@ import { SharedCredentialsRepository } from '@/databases/repositories/shared-cre
 import * as Db from '@/db';
 import { EventService } from '@/events/event.service';
 import { ExternalHooks } from '@/external-hooks';
-import type { ICredentialsDb } from '@/interfaces';
 import type { CredentialRequest } from '@/requests';
+import type { ICredentialsDb } from '@/types-db';
 
 import type { IDependency, IJsonSchema } from '../../../types';
