+++ packages/cli/src/services/annotation-tag.service.ee.ts
@@ -3,7 +3,7 @@ import { Service } from '@n8n/di';
 import type { AnnotationTagEntity } from '@/databases/entities/annotation-tag-entity.ee';
 import { AnnotationTagRepository } from '@/databases/repositories/annotation-tag.repository.ee';
 import { validateEntity } from '@/generic-helpers';
-import type { IAnnotationTagDb, IAnnotationTagWithCountDb } from '@/interfaces';
+import type { IAnnotationTagDb, IAnnotationTagWithCountDb } from '@/types-db';
 
 type GetAllResult<T> = T extends { withUsageCount: true }
 	? IAnnotationTagWithCountDb[]
