+++ packages/cli/src/services/tag.service.ts
@@ -4,7 +4,7 @@ import type { TagEntity } from '@/databases/entities/tag-entity';
 import { TagRepository } from '@/databases/repositories/tag.repository';
 import { ExternalHooks } from '@/external-hooks';
 import { validateEntity } from '@/generic-helpers';
-import type { ITagWithCountDb } from '@/interfaces';
+import type { ITagWithCountDb } from '@/types-db';
 
 type GetAllResult<T> = T extends { withUsageCount: true } ? ITagWithCountDb[] : TagEntity[];
