+++ packages/cli/src/services/user.service.ts
@@ -9,10 +9,11 @@ import { User } from '@/databases/entities/user';
 import { UserRepository } from '@/databases/repositories/user.repository';
 import { InternalServerError } from '@/errors/response-errors/internal-server.error';
 import { EventService } from '@/events/event.service';
-import type { Invitation, PublicUser } from '@/interfaces';
+import type { Invitation } from '@/interfaces';
 import type { PostHogClient } from '@/posthog';
 import type { UserRequest } from '@/requests';
 import { UrlService } from '@/services/url.service';
+import type { PublicUser } from '@/types-db';
 import { UserManagementMailer } from '@/user-management/email';
 
 import { PublicApiKeyService } from './public-api-key.service';
