+++ packages/cli/src/webhooks/__tests__/waiting-forms.test.ts
@@ -3,9 +3,9 @@ import { mock } from 'jest-mock-extended';
 import { FORM_NODE_TYPE, WAITING_FORMS_EXECUTION_STATUS, type Workflow } from 'n8n-workflow';
 
 import type { ExecutionRepository } from '@/databases/repositories/execution.repository';
+import type { IExecutionResponse } from '@/types-db';
 import { WaitingForms } from '@/webhooks/waiting-forms';
 
-import type { IExecutionResponse } from '../../interfaces';
 import type { WaitingWebhookRequest } from '../webhook.types';
 
 describe('WaitingForms', () => {
