+++ packages/cli/src/webhooks/waiting-webhooks.ts
@@ -13,8 +13,8 @@ import {
 import { ExecutionRepository } from '@/databases/repositories/execution.repository';
 import { ConflictError } from '@/errors/response-errors/conflict.error';
 import { NotFoundError } from '@/errors/response-errors/not-found.error';
-import type { IExecutionResponse } from '@/interfaces';
 import { NodeTypes } from '@/node-types';
+import type { IExecutionResponse } from '@/types-db';
 import * as WebhookHelpers from '@/webhooks/webhook-helpers';
 import * as WorkflowExecuteAdditionalData from '@/workflow-execute-additional-data';
