+++ packages/cli/test/integration/shared/types.ts
@@ -7,7 +7,7 @@ import type TestAgent from 'supertest/lib/agent';
 import type { CredentialsEntity } from '@/databases/entities/credentials-entity';
 import type { Project } from '@/databases/entities/project';
 import type { User } from '@/databases/entities/user';
-import type { ICredentialsDb } from '@/interfaces';
+import type { ICredentialsDb } from '@/types-db';
 
 import type { LicenseMocker } from './license';
